+++ disk/src/android_sparse.rs
@@ -281,7 +281,7 @@ impl FileReadWriteAtVolatile for AndroidSparse {
         match chunk {
             Chunk::DontCare => {
                 subslice.write_bytes(0);
-                Ok(subslice.size() as usize)
+                Ok(subslice.size())
             }
             Chunk::Raw(file_offset) => self
                 .file
@@ -293,10 +293,10 @@ impl FileReadWriteAtVolatile for AndroidSparse {
                     .cloned()
                     .cycle()
                     .skip(chunk_offset_mod as usize)
-                    .take(subslice.size() as usize)
+                    .take(subslice.size())
                     .collect();
                 subslice.copy_from(&filled_memory);
-                Ok(subslice.size() as usize)
+                Ok(subslice.size())
             }
         }
     }
